Fixed Issues + Comments
Added comments and text overlapping game

+++ SnakeGame.java
@@ -45,6 +45,7 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
     boolean v3 = false;
     int delay = 100;
     Color randomColor = Color.green;
+    Color randomColorKeep = Color.white;
     int r = 0;
     int g = 0;
     int b = 0;
@@ -61,31 +62,31 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
         checkVersion(version);
         if(v1)
             randomize();
-
+        //Game objects
         snakeHead = new Tile(5,5);
         snakeBody = new ArrayList<Tile>();
         food = new Tile(10,100);
         badfood = new Tile(10,101);
         instantfood = new Tile(10,102);
+        //Tools
         random = new Random();
+        //Options
         Easy = new Tile(10,5);
         Medium = new Tile(10,10);
         Hard = new Tile(10,15);
         color = new Tile(20,5);
         
         
-
+        //Speed taken
         if(v2)
             delay = 50;
         if(v3)
             delay = 200;
         velocityX = 0;
         velocityY =0;
-        /* 
-        gameLoop = new Timer(delay, this);
-        gameLoop.start();
-         */
+        
     }
+    //Random Color Given, Optimized in mouse clicks
     public void randomize(){
         r = (int)(Math.random()*155)+100;
         g = (int)(Math.random()*155)+100;
@@ -93,21 +94,21 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
         randomColor = new Color(r,g,b);
         
         if(r == 0){
-            System.out.println("Orange chosen");
+            //System.out.println("Orange chosen");
             randomColor = Color.orange;
         }
         if(r == 1){
             randomColor = Color.pink;
-            System.out.println("Pink chosen");
+            //System.out.println("Pink chosen");
 
         }
         if(r == 2){
-            System.out.println("Cyan chosen");
+            //System.out.println("Cyan chosen");
             randomColor = Color.cyan;
         }
         
     }
-
+    //Frame paint (Game)
     public void paintComponent(Graphics g){
         super.paintComponent(g);
         draw(g);
@@ -148,57 +149,72 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
         else{
             g.drawString("Score: " + String.valueOf(snakeBody.size()), tileSize -16, tileSize);
         }
-
+        //Filling in the rectangle colors
         g.setColor(Color.red);
         g.fillRect(Easy.x*tileSize,Easy.y*tileSize, tileSize, tileSize);
         g.setColor(Color.blue);
         g.fillRect(Medium.x*tileSize, Medium.y*tileSize, tileSize, tileSize);
         g.setColor(Color.orange);
         g.fillRect(Hard.x*tileSize, Hard.y*tileSize, tileSize, tileSize);
-        g.setColor(Color.white);
+        g.setColor(randomColorKeep);
         g.fillRect(color.x*tileSize,color.y*tileSize,tileSize,tileSize);
     
     
             //Score
             g.setColor(Color.white);
             g.setFont(new Font("Arial",Font.PLAIN,15));
-            //g.setFont(new Font("Arial", Font.PLAIN, 15))
-            g.drawString("Snake Game",boardWidth/2  - 42, tileSize);
-            g.drawString("The green square is you the \"the player\"",boardWidth/3 - 15, tileSize+30);
-            g.drawString("The red square is food to grow",boardWidth/3 + 10, tileSize +48);
-            g.drawString("The blue square will shrink you",boardWidth/3 + 10, tileSize + 66);
-            g.drawString("The orange square will kill you",boardWidth/3 + 10, tileSize + 84 );
+            //Adding text till the game mode is chosen
+            if(!v1 && !v2 && !v3){
+                //info
+                g.drawString("Snake Game",boardWidth/2  - 42, tileSize);
+                g.drawString("The green square is you the \"the player\"",boardWidth/3 - 15, tileSize+30);
+                g.drawString("The red square is food to grow",boardWidth/3 + 10, tileSize +48);
+                g.drawString("The blue square will shrink you",boardWidth/3 + 10, tileSize + 66);
+                g.drawString("The orange square will kill you",boardWidth/3 + 10, tileSize + 84 );
+                //Position of text to reference the buttons
+                g.drawString("Easy Mode", 250+45,125+18);
+                g.drawString("Medium Mode", 250+45, 250+18);
+                g.drawString("Hard Mode", 250+45, 375+18);
+                g.drawString("Random Color", 470, 168);
+                g.drawString("Change", 485,186);
+            }
         }
-    
+    //Food random
     public void placeFood(){
         food.x = random.nextInt(boardWidth/tileSize);
         food.y = random.nextInt(boardWidth/tileSize);
     }
-
+    //Shrinking Food
     public void placeBadFood(){
         badfood.x=random.nextInt(boardWidth/tileSize);
         badfood.y=random.nextInt(boardWidth/tileSize);
     }
+    //Death Food
     public void placeInstantFood(){
         instantfood.x=random.nextInt(boardWidth/tileSize);
         instantfood.y=random.nextInt(boardWidth/tileSize);
     }
+    //Colliding
     public boolean collision(Tile tile1, Tile tile2){
         return tile1.x == tile2.x && tile1.y == tile2.y;
     }
+    //Snake movement
     public void move(){
         //get Food
         if(collision(snakeHead, food)){
             snakeBody.add(new Tile(food.x, food.y));
             placeFood();
+            //If medium
             if(v2 == true){
                 placeBadFood();
             }
+            //If hard
             if(v3 == true){
                 placeBadFood();
                 placeInstantFood();
             }
         }
+        //Shrink food detection
         if(collision(snakeHead, badfood)){
             if (snakeBody.size() == 0) {
                 // Game over if snake's length is 1
@@ -208,6 +224,7 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
             }
             placeBadFood();
         }
+        //Death food detection
         if(collision(snakeHead,instantfood)){
             gameOver=true;
         }
@@ -236,6 +253,7 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
                 gameOver =true;
             }
         }
+        //Checking for borders
         if (snakeHead.x*tileSize<0|| snakeHead.x*tileSize >boardWidth||
             snakeHead.y*tileSize <0 || snakeHead.y*tileSize >boardHeight){
             gameOver =true;
@@ -249,6 +267,7 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
         if(v==3)
             v3=true;
     }
+    //Game running
     @Override
     public void actionPerformed(ActionEvent e){
         move();
@@ -257,7 +276,7 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
             gameLoop.stop();
         }
     }
-    
+    //Snake movement
     @Override
     public void keyPressed(KeyEvent e) {
         if (e.getKeyCode() == KeyEvent.VK_UP && velocityY != 1){
@@ -287,12 +306,15 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
             randomColor = new Color(r,g,b);
         }
     }
+    //Picking gamemode
     @Override
     public void mouseClicked(MouseEvent e)
-{
+    {
     int mouseX=e.getX();
     int mouseY=e.getY();
-
+    //Pick decision mouse
+    //First 4 lines of each if repositions used assets outside out of mouse click range
+    //Next lines respective to difficulty, game starts
     if (mouseX >= Easy.x*tileSize && mouseX <(Easy.x+1)*tileSize && mouseY >= Easy.y*tileSize && mouseY < (Easy.y+1) *tileSize){
         Easy = new Tile(100,100);
         Medium = new Tile (100, 100);
@@ -334,8 +356,10 @@ public class SnakeGame extends JPanel implements ActionListener,KeyListener,Mous
         g = (int)(Math.random()*155)+100;
         b = (int)(Math.random()*155)+100;
         randomColor = new Color(r,g,b);
+        randomColorKeep = new Color(r,g,b);
     }
-}    //Do Not Need
+}    
+    //NOT NEEDED, DO NOT MODIFY
     @Override
     public void keyTyped(KeyEvent e) {
     }
